Move get_selected_values and related helpers to utils
