+++ app/apps/parallel_benchmarks.py
@@ -2,7 +2,6 @@ from re import U, split, sub
 import streamlit as st
 import numpy as np
 import pandas as pd
-from functools import reduce
 
 from nested_dict import nested_dict
 from pprint import pprint
@@ -14,7 +13,7 @@ import pandas as pd
 import pandas.io.json as pdjson
 import seaborn as sns
 from apps import benchstruct
-from apps.utils import fmt_variant, unfmt_variant
+from apps.utils import get_selected_values, flatten
 
 
 def app():
@@ -32,62 +31,11 @@ def app():
     st.header("Select variants")
     n = int(st.text_input("Number of variants", "1", key=benches.config["bench_type"]))
 
-    containers = [st.columns(3) for i in range(n)]
-
-    # [[a11, a12 ... a1n], [a21, a22 ... a2n], ... [am1, am2 ... amn]] => [a11]
-    def flatten(lst):
-        return reduce(lambda a, b: a + b, lst)
-
-    # [(a1, b1), (a2, b2) ... (an, bn)] => ([a1, a2, ... an], [b1, b2, ... bn])
-    def unzip(lst):
-        return list(zip(*lst))
-
-    def unzip_dict(d):
-        a = unzip(list(d))
-        # st.write(a)
-        commit_variant_tuple_lst = [(x1, x2) for x1, x2 in zip(a[0], a[1])]
-        return commit_variant_tuple_lst
-
-    def get_selected_values(n):
-        lst = []
-        for i in range(n):
-            # create the selectbox in columns
-            host_val = containers[i][0].selectbox(
-                "hostname", benches.structure.keys(), key=str(i) + "0_sequential"
-            )
-            timestamp_val = containers[i][1].selectbox(
-                "timestamp",
-                benches.structure[host_val].keys(),
-                key=str(i) + "1_sequential",
-            )
-            commit_variant_tuple_lst = unzip_dict(
-                (benches.structure[host_val][timestamp_val]).items()
-            )
-            # st.write(commit_variant_tuple_lst)
-            fmtted_variants = [fmt_variant(c, v) for c, v in commit_variant_tuple_lst]
-            fmtted_variants = flatten(fmtted_variants)
-            # st.write(fmtted_variants)
-            variant_val = containers[i][2].selectbox(
-                "variant", fmtted_variants, key=str(i) + "2_sequential"
-            )
-            selected_commit, selected_variant = unfmt_variant(variant_val)
-            lst.append(
-                {
-                    "host": host_val,
-                    "timestamp": timestamp_val,
-                    "commit": selected_commit,
-                    "variant": selected_variant,
-                }
-            )
-        return lst
-
     selected_benches = benchstruct.BenchStruct(
         "parallel", artifacts_dir, "_1.orunchrt.summary.bench"
     )
-    _ = [
+    for f in get_selected_values(n, benches):
         selected_benches.add(f["host"], f["timestamp"], f["commit"], f["variant"])
-        for f in get_selected_values(n)
-    ]
     selected_benches.sort()
 
     # Expander for showing bench files
